Split CallLinker into static and dynamic part
